Move and rename `ByteCostConfigBuilder`
